Fix PR requests
