+++ front/src/components/Actu/DeclarationDialogs/ConsistencyErrorsDialog.js
@@ -6,12 +6,9 @@ import DialogContentText from '@material-ui/core/DialogContentText'
 import CustomColorButton from '../../Generic/CustomColorButton'
 import CustomDialog from '../../Generic/CustomDialog'
 
-const ConsistencyErrorsDialogs = ({
-  consistencyErrors,
-  onCancel,
-  confirmAndIgnoreErrors,
-  props,
-}) => {
+const ConsistencyErrorsDialogs = (props) => {
+  const { consistencyErrors, onCancel, confirmAndIgnoreErrors } = props
+
   return (
     <CustomDialog
       content={
